+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
+> NOTE: This release has **non-backwards compatible** changes.
+### Changed
+- Database schema has been changed in non-backwards compatible way. 
 
 ## 2.7.3 - 2022-04-06